requirements.txt
scipy==1.6.1
numpy==1.19.2
matplotlib==3.3.2
mpi4py==3.0.3
pyyaml==5.3.1
